Use stride in elements
This commit replaces strides in bytes with strides in elements. This is achieved by relying on the PitchedElement concept which constrains the element type of image and image view. The bytes accessor is not needed anymore, but we still need a "located" accessor, i.e. the default accessor with a MemoryLocation embedded.